adjust speed, adjust pipe spacing, rename xmover to pipemover for clarity, use floats for position and speed

+++ lib/flappy_elixir/components/x_position.ex
@@ -3,5 +3,5 @@ defmodule FlappyElixir.Components.XPosition do
   Documentation for XPosition components.
   """
   use ECSx.Component,
-    value: :integer
+    value: :float
 end
